requirements.txt
numpy==2.0.2
pandas==2.2.3
tensorflow==2.18.0

# tests
httpx           # HTTP client
pytest-asyncio  # asynchronous I/O support for pytest

# img
pillow

#models
python-dotenv
python-multipart

# web/API
streamlit
requests
uvicorn         # web server
fastapi
